Split `tests/P1206R7_ranges_to` into four parts
,... to reduce resource consumption. The obvious approach here is to split the three big pieces (the `sequence_instantiator`, `mappish_instantiator`, and `settish_instantiator` tests) each into their own tests, leaving behind the three small pieces (`test_reservable`, `test_common_constructible`, and `test_nested_range`).

Fun fact: there are now 50 `P1206R7` tests in the `std` suite.
